pyproject.toml
[project]
name = "marketplace-andes-backend"
version = "0.0.1"
description = "Marketplace Andes"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "passlib[argon2]>=1.7.4",
    "psycopg[binary]>=3.3.3",
    "pyjwt>=2.12.1",
    "sqlmodel>=0.0.37",
]

[project.optional-dependencies]
otel = [
    "opentelemetry-exporter-otlp-proto-grpc>=1.40.0",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
    "opentelemetry-instrumentation-logging>=0.60b1",
    "opentelemetry-instrumentation-psycopg>=0.60b1",
    "opentelemetry-instrumentation-sqlalchemy>=0.60b1",
]

[tool.fastapi]
entrypoint = "src.marketplace_andes_backend.app:app"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.6",
    "testcontainers[postgres]>=4.14.2",
    "ty>=0.0.23",
]
migrations = ["alembic>=1.18.4", "sqlmodel>=0.0.37", "psycopg[binary]>=3.3.3"]
viz = ["duckdb>=1.3.2", "matplotlib>=3.10.6", "pandas>=2.3.2"]

[tool.alembic]
script_location = "%(here)s/alembic"
prepend_sys_path = ["."]
output_encoding = "utf-8"

[[tool.alembic.post_write_hooks]]
name = "ruff"
type = "module"
module = "ruff"
options = "check --fix REVISION_SCRIPT_FILENAME"
